requirements.txt
Django>=3.1.5,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.8.6,<2.9.0
flake8>=3.8.4,<3.9.0
isort>=5.7.0
coverage>=5.4,<5.5
coverage-badge>=1.0.1